Direction and sleep now work properly. Split functions into separate files
