+++ Src/gpio.h
@@ -1,0 +1,8 @@
+#include "stm32f0xx_hal.h"
+
+void gpio_output_init(GPIO_TypeDef * port, uint32_t pin);
+void gpio_alternate_init(GPIO_TypeDef * port, uint32_t pin);
+void gpio_input_init(GPIO_TypeDef * port, uint32_t pin);
+void gpio_write_reg32(__IO uint32_t * reg, uint32_t pin, uint32_t bits);
+void gpio_write_reg16(__IO uint32_t * reg, uint32_t pin, uint32_t bits);
+void gpio_toggle_reg16(__IO uint32_t * reg, uint32_t pin);
