+++ Src/stepper.h
@@ -1,0 +1,52 @@
+// stepper PWM
+#define PRESCALE 50
+#define AUTO_RELOAD 20
+#define DUTY_CYCLE 2
+
+// stepper  step & gear constants
+#define GEAR_C             30     // approx gear circumference, in mm
+#define STEP_ANGLE         0.0314 // radians, approx 1.8*
+#define GEAR_RADIUS        4.75   // mm
+#define LINEAR_DISTANCE    0.15   // mm, STEP_ANGLE*GEAR_RADIUS
+#define STEPS_PER_REV_FULL 200    // full step
+#define STEPS_PER_REV_16   3200   // 1/16 microstep
+#define STEPS_PER_MM_FULL  7      //  (STEPS_PER_REV_FULL*MICROSTEP)/GEAR_C
+#define STEPS_PER_MM_16    107    // 1/16 microstep
+
+// axis
+#define X 0
+#define Y 1
+#define OFF -1
+#define CW   0 // direction
+#define CCW  1
+
+// axis pins (GPIOC)
+#define X_DIR   10
+#define X_RESET 12
+#define X_CAL   4 // input
+#define X_SLEEP 6
+#define X_STEP  8
+#define X_PWM   TIM3->CCR3
+
+#define Y_DIR   11
+#define Y_RESET 3
+#define Y_CAL   5 // input
+#define Y_SLEEP 7
+#define Y_STEP  9
+#define Y_PWM   TIM3->CCR4
+
+// helper macros
+#define SLEEP(axis) (GPIOC->ODR &= (axis == X) ? ~(1 << X_SLEEP) : ~(1 << Y_SLEEP))
+#define WAKE(axis)  (GPIOC->ODR |= (axis == X) ?  (1 << X_SLEEP) :  (1 << Y_SLEEP))
+
+// chess specific
+#define SQUARE_WIDTH 51 // mm
+void step_squares(int axis, int n, int dir);
+
+// func definitions
+void step_init(void);
+void step_stop(int axis);
+void step(void);
+void stepn(int axis, int n, int dir);
+void step_mm(int axis, int mm, int dir);
+void set_dir(int axis, int dir);
